# refactor: faster html escape