Removing the need for regex yields a ~240ops/sec increase.

Before:
```
  pug                x 2,121 ops/sec ±1.50% (91 runs sampled)
  handlebars         x 674 ops/sec ±1.39% (91 runs sampled)
  ejs                x 327 ops/sec ±1.57% (87 runs sampled)
  yeahjs             x 884 ops/sec ±1.30% (91 runs sampled)
  dot                x 608 ops/sec ±2.21% (91 runs sampled)
  art-template       x 2,085 ops/sec ±0.85% (93 runs sampled)
  tempura            x 2,186 ops/sec ±1.64% (93 runs sampled)
```
After:
```
  pug                x 2,010 ops/sec ±2.03% (89 runs sampled)
  handlebars         x 674 ops/sec ±1.30% (91 runs sampled)
  ejs                x 319 ops/sec ±1.61% (89 runs sampled)
  yeahjs             x 883 ops/sec ±1.19% (92 runs sampled)
  dot                x 617 ops/sec ±1.59% (91 runs sampled)
  art-template       x 2,070 ops/sec ±1.10% (93 runs sampled)
  tempura            x 2,426 ops/sec ±1.65% (92 runs sampled)
```