@intrnl 2022-07-24
It's faster when involving values that needs to actually be escaped, but it can be slower when things don't need to actually be escaped.